Code refactor (#1501)
* Code refactor

* Incremented patch version

* Code refactor

---------
