+++ src/Compiler/Constraints.h
@@ -85,30 +85,6 @@ class Constraints {
   const std::vector<OBJECT_PROPERTY> get_property();
   nlohmann::json get_property_by_json();
   void write_property(const std::string& filepath);
-  std::map<std::string, std::string>& getInputOutputMap() {
-    return m_input_output_map;
-  }
-  std::map<std::string, std::string>& getOutputInputMap() {
-    return m_output_input_map;
-  }
-
-  std::string PIO2InnerNet(const std::string& orig);
-  std::string InnerNet2PIO(const std::string& orig);
-
-  std::string FindAliasInInputOutputMap(const std::string& orig);
-  void ComputePrimaryMaps();
-  std::map<std::string, std::string>& getPrimaryInputMap() {
-    return m_primary_input_map;
-  }
-  std::map<std::string, std::string>& getPrimaryOutputMap() {
-    return m_primary_output_map;
-  }
-  std::map<std::string, std::string>& getReversePrimaryInputMap() {
-    return m_reverse_primary_input_map;
-  }
-  std::map<std::string, std::string>& getReversePrimaryOutputMap() {
-    return m_reverse_primary_output_map;
-  }
 
  protected:
   Compiler* m_compiler = nullptr;
@@ -121,12 +97,6 @@ class Constraints {
   std::map<std::string, float> m_clockPeriodMap;
   std::vector<OBJECT_PROPERTY> m_object_properties;
   ConstraintPolicy m_constraintPolicy = ConstraintPolicy::SDCCompatible;
-  std::map<std::string, std::string> m_input_output_map;
-  std::map<std::string, std::string> m_output_input_map;
-  std::map<std::string, std::string> m_primary_input_map;
-  std::map<std::string, std::string> m_primary_output_map;
-  std::map<std::string, std::string> m_reverse_primary_input_map;
-  std::map<std::string, std::string> m_reverse_primary_output_map;
 };
 
 }  // namespace FOEDAG
