+++ src/DesignQuery/DesignQuery.cpp
@@ -1071,7 +1071,8 @@ bool DesignQuery::RegisterCommands(TclInterpreter* interp, bool batchMode) {
       for (int i = 1; i < argc; i++) {
         std::string arg = argv[i];
         std::string tmp = StringUtils::replaceAll(arg, "@*@", "{*}");
-        tmp = constraints->PIO2InnerNet(tmp);
+        tmp =
+            designQuery->GetCompiler()->getNetlistEditData()->PIO2InnerNet(tmp);
         if (tmp != "{*}") constraints->addKeep(tmp);
         arguments.push_back(tmp);
       }
